+++ R/utils-misc.r
@@ -9,7 +9,7 @@ assert_suggested <- function(package) {
 }
 
 is_utc <- function(tz) {
-    (tz %||% Sys.timezone()) %in% c("UTC", "GMT", "Etc/UTC", "Etc/GMT", "GMT-0", "GMT+0", "GMT0")
+    tz %in% c("UTC", "GMT", "Etc/UTC", "Etc/GMT", "GMT-0", "GMT+0", "GMT0")
 }
 
 `%||%` <- function(x, y) if (is.null(x)) y else x